+++ app/javascript/packs/application.js
@@ -37,13 +37,13 @@ import '../Codemirror/LiquidMultiplexModes';
 
 import ConventionAdmin from '../ConventionAdmin';
 import DatagridFilters from '../Datagrid/DatagridFilters';
-import EditUserConProfile from '../UserConProfileAdmin/EditUserConProfile';
+import EditUserConProfile from '../UserConProfiles/EditUserConProfile';
 import EventAdmin from '../EventAdmin';
 import EventProposalForm from '../FormPresenter/RESTForms/EventProposalForm';
 import EventListSortDropdown from '../UIComponents/EventListSortDropdown';
 import FormJSONEditor from '../FormJSONEditor';
 import LiquidHTMLEditor from '../StandaloneFormControls/LiquidHTMLEditor';
-import MyProfileForm from '../UserConProfileAdmin/MyProfileForm';
+import MyProfileForm from '../UserConProfiles/MyProfileForm';
 import PageAdminDropdown from '../UIComponents/PageAdminDropdown';
 import RoomsAdmin from '../RoomsAdmin';
 import ScheduleGrid from '../ScheduleGrid';
